--- Source/TrashPanda/MyAnimInstance.cpp
@@ -1,8 +1,0 @@
-// All Rights Reserved for Students Graduating TFS Summer 2017
-
-#include "TrashPanda.h"
-#include "MyAnimInstance.h"
-
-
-
-
