--- Source/TrashPanda/MyAnimInstance.h
@@ -1,19 +1,0 @@
-// All Rights Reserved for Students Graduating TFS Summer 2017
-
-#pragma once
-
-#include "Animation/AnimInstance.h"
-#include "MyAnimInstance.generated.h"
-
-/**
- * 
- */
-UCLASS()
-class TRASHPANDA_API UMyAnimInstance : public UAnimInstance
-{
-	GENERATED_BODY()
-	
-	
-	
-	
-};
