+++ Source/TrashPanda/Player/Chip.h
@@ -34,13 +34,27 @@ public:
 	bool GetIsLightAttacking();
 	bool GetIsHeavyAttacking();
 
-protected:
 
-	int32 Health;
-	int32 MaxHealth;
-	int32 Fury;
+	float CurrentHealth;
+	float MaxHealth;
+	int32 CurrentFury;
 	int32 MaxFury;
 
+	float GetHealthAsPercentage();
+	float GetHealth();
+	float GetMaxHealth();
+	float GetFuryAsPercentage();
+	float GetFury();
+	float GetMaxFury();
+	int GetHConsumablesQuantity();
+	int GetFConsumablesQuantity();
+
+	//Functions to debug health and fury
+	void DebugHealth();
+	void DebugFury();
+
+
+protected:
 	bool bisRabid;
 	bool bisLightAttacking;
 	bool bisHeavyAttacking;
@@ -77,11 +91,14 @@ protected:
 	class UInventoryWidget* InvWidget;
 
 	UPROPERTY(EditDefaultsOnly)
+	TSubclassOf<class UChipHUDWidget>ChipHUDWidgetClass;
+	
+	class UChipHUDWidget* ChipHUDWidget;
+
+
 	TSubclassOf<class UCharacterWidgetSwitcher> WidgetSwitcherClass;
 
 	class UCharacterWidgetSwitcher* SwitchWidget;
-
-
 
 	UPROPERTY(VisibleAnywhere, Category = Player)
 	TArray<AActor*> itemsInRange;
@@ -102,6 +119,8 @@ protected:
 	void Death();
 
 	void ReadInv();
+
+
 
 	UFUNCTION()
 	void OnOverlapBegin(UPrimitiveComponent* OverlappedComponent, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult);
